Add support for persisting hdive through transitions
This is a lot easier than doing the same for vertical cdash, because we can piggyback off the code for cdash through transitions.

It's not totally perfect - sometimes the knight's hitbox is higher than it should be, and sometimes it isn't (I've only had this when diving right though) - but it seems to work, for the most part. for now.